+++ AAEmu.Game/Core/Network/Connections/GameConnection.cs
@@ -28,7 +28,7 @@ namespace AAEmu.Game.Core.Network.Connections
         private Session _session;
 
         public uint Id => _session.Id;
-        public uint AccountId { get; set; }
+        public ulong AccountId { get; set; }
         public IPAddress Ip => _session.Ip;
         public PacketStream LastPacket { get; set; }
